Move SendSlackMessages to Slack Namespace
