+++ dependapanda.rb
@@ -2,14 +2,14 @@ require_relative 'lib/loader'
 
 class Dependapanda
   def send_simple_message
-    UseCases::SendSlackMessages.new(
+    UseCases::Slack::SendMessages.new(
       message_presenter: Presenters::Slack::SimpleMessage.new,
       scheduler: UseCases::Slack::Schedulers::Weekday.new
     ).execute
   end
 
   def send_full_message
-    UseCases::SendSlackMessages.new(
+    UseCases::Slack::SendMessages.new(
       message_presenter: Presenters::Slack::FullMessage.new,
       scheduler: UseCases::Slack::Schedulers::Weekday.new
     ).execute
