+++ lib/loader.rb
@@ -7,12 +7,12 @@ require_relative './use_cases/cache'
 require_relative './use_cases/fetch_pull_requests'
 require_relative './use_cases/fetch_pull_request_count'
 require_relative './use_cases/fetch_teams'
-require_relative './use_cases/send_slack_messages'
 require_relative './use_cases/group_pull_requests_by_application'
 require_relative './use_cases/group_applications_by_team'
 require_relative './use_cases/group_pull_requests_by_gem'
 require_relative './use_cases/slack/schedulers/every_day'
 require_relative './use_cases/slack/schedulers/weekday'
+require_relative './use_cases/slack/send_messages'
 require_relative './use_cases/fetch_gemfiles'
 require_relative './use_cases/save_gemfiles'
 require_relative './gateways/pull_request'
